+++ mlxtend/classifier/stacking_classification.py
@@ -13,6 +13,7 @@ import warnings
 import numpy as np
 from scipy import sparse
 from sklearn.base import TransformerMixin, clone
+from sklearn.preprocessing import LabelEncoder
 
 from ..externals.estimator_checks import check_is_fitted
 from ..externals.name_estimators import _name_estimators
@@ -95,6 +96,9 @@ class StackingClassifier(_BaseXComposition, _BaseStackingClassifier, Transformer
         Fitted classifiers (clones of the original classifiers)
     meta_clf_ : estimator
         Fitted meta-classifier (clone of the original meta-estimator)
+    classes_ : ndarray of shape (n_classes,) or list of ndarray if `y` \
+            is of type `"multilabel-indicator"`.
+            Class labels.
     train_meta_features : numpy array, shape = [n_samples, n_classifiers]
         meta-features for training data, where n_samples is the
         number of samples
@@ -174,6 +178,13 @@ class StackingClassifier(_BaseXComposition, _BaseStackingClassifier, Transformer
         else:
             self.clfs_ = self.classifiers
             self.meta_clf_ = self.meta_classifier
+
+        if y.ndim > 1:
+            self._label_encoder = [LabelEncoder().fit(yk) for yk in y.T]
+            self.classes_ = [le.classes_ for le in self._label_encoder]
+        else:
+            self._label_encoder = LabelEncoder().fit(y)
+            self.classes_ = self._label_encoder.classes_
 
         if self.fit_base_estimators:
             if self.verbose > 0: